Fixing circular dependencies from field-space, space-field and the ast directory.
