+++ packages/malloy/src/lang/ast/ast-main.ts
@@ -22,6 +22,8 @@
  */
 import { cloneDeep } from "lodash";
 import * as model from "../../model/malloy_types";
+import { Dialect } from "../../dialect/dialect";
+import { getDialect } from "../../dialect/dialect_map";
 import {
   AtomicFieldType,
   DivFragment,
@@ -39,17 +41,6 @@ import {
   TimestampUnit,
 } from "../../model/malloy_types";
 import { Segment as ModelQuerySegment } from "../../model/malloy_query";
-import {
-  DefSpace,
-  FieldSpace,
-  ReduceFieldSpace,
-  ProjectFieldSpace,
-  DynamicSpace,
-  QuerySpace,
-  IndexFieldSpace,
-  LookupResult,
-  ResultSpace,
-} from "../field-space";
 import { MessageLogger } from "../parse-log";
 import { MalloyTranslation, ModelDataRequest } from "../parse-malloy";
 import {
@@ -66,7 +57,6 @@ import {
   isEquality,
   isGranularResult,
 } from "./ast-types";
-import { QueryField } from "../space-field";
 import { makeSQLBlock } from "../../model/sql_block";
 import { inspect } from "util";
 import {
@@ -76,6 +66,7 @@ import {
   timeResult,
 } from "./time-utils";
 import { nullsafeNot } from "./apply-expr";
+import { mergeFields, nameOf } from "../field-utils";
 
 /*
  ** For times when there is a code generation error but your function needs
@@ -3274,5 +3265,917 @@ export class Range extends ExpressionDef {
 
   getExpression(_fs: FieldSpace): ExprValue {
     return errorFor("a range is not a value");
+  }
+}
+
+/**
+ * A FieldSpace is a hierarchy of namespaces, where the leaf nodes
+ * are fields. A FieldSpace can lookup fields, and generate a StructDef
+ */
+export interface FieldSpace {
+  type: "fieldSpace";
+  structDef(): model.StructDef;
+  emptyStructDef(): model.StructDef;
+  lookup(symbol: FieldName[]): LookupResult;
+  dialectObj(): Dialect | undefined;
+  whenComplete: (step: () => void) => void;
+}
+
+/**
+ * Used to detect references to fields in the statement which defines them
+ */
+export class DefSpace implements FieldSpace {
+  readonly type = "fieldSpace";
+  foundCircle = false;
+  constructor(
+    readonly realFS: FieldSpace,
+    readonly circular: FieldDeclaration
+  ) {}
+  structDef(): model.StructDef {
+    return this.realFS.structDef();
+  }
+  emptyStructDef(): model.StructDef {
+    return this.realFS.emptyStructDef();
+  }
+  lookup(symbol: FieldName[]): LookupResult {
+    if (symbol[0] && symbol[0].refString === this.circular.defineName) {
+      this.foundCircle = true;
+      return {
+        error: `Circular reference to '${this.circular.defineName}' in definition`,
+        found: undefined,
+      };
+    }
+    return this.realFS.lookup(symbol);
+  }
+  dialectObj(): Dialect | undefined {
+    return this.realFS.dialectObj();
+  }
+  whenComplete(step: () => void): void {
+    this.realFS.whenComplete(step);
+  }
+}
+
+export class StaticSpace implements FieldSpace {
+  readonly type = "fieldSpace";
+  private memoMap?: FieldMap;
+  protected fromStruct: model.StructDef;
+
+  constructor(sourceStructDef: model.StructDef) {
+    this.fromStruct = sourceStructDef;
+  }
+
+  whenComplete(step: () => void): void {
+    step();
+  }
+
+  dialectObj(): Dialect | undefined {
+    try {
+      return getDialect(this.fromStruct.dialect);
+    } catch {
+      return undefined;
+    }
+  }
+
+  fromFieldDef(from: model.FieldDef): SpaceField {
+    if (from.type === "struct") {
+      return new StructSpaceField(from);
+    } else if (model.isTurtleDef(from)) {
+      return new QueryFieldStruct(this, from);
+    }
+    // TODO field has a filter and needs to be a filteredalias
+    return new ColumnSpaceField(from);
+  }
+
+  private get map(): FieldMap {
+    if (this.memoMap === undefined) {
+      this.memoMap = {};
+      for (const f of this.fromStruct.fields) {
+        const name = f.as || f.name;
+        this.memoMap[name] = this.fromFieldDef(f);
+      }
+      if (this.fromStruct.parameters) {
+        for (const [paramName, paramDef] of Object.entries(
+          this.fromStruct.parameters
+        )) {
+          if (this.memoMap[paramName]) {
+            throw new Error(
+              `In struct '${this.fromStruct.as || this.fromStruct.name}': ` +
+                ` : Field and parameter name conflict '${paramName}`
+            );
+          }
+          this.memoMap[paramName] = new DefinedParameter(paramDef);
+        }
+      }
+    }
+    return this.memoMap;
+  }
+
+  protected dropEntries(): void {
+    this.memoMap = {};
+  }
+
+  protected dropEntry(name: string): void {
+    delete this.map[name];
+  }
+
+  protected entry(name: string): SpaceEntry | undefined {
+    return this.map[name];
+  }
+
+  protected setEntry(name: string, value: SpaceEntry): void {
+    this.map[name] = value;
+  }
+
+  protected entries(): [string, SpaceEntry][] {
+    return Object.entries(this.map);
+  }
+
+  structDef(): model.StructDef {
+    return this.fromStruct;
+  }
+
+  emptyStructDef(): model.StructDef {
+    return { ...this.fromStruct, fields: [] };
+  }
+
+  lookup(path: FieldName[]): LookupResult {
+    const head = path[0];
+    const rest = path.slice(1);
+    const found = this.entry(head.refString);
+    if (!found) {
+      return { error: `'${head}' is not defined`, found };
+    }
+    if (found instanceof SpaceField) {
+      /*
+       * TODO cache defs, post the addReference call to whenComplete
+       *
+       * In the cleanup phase of query space construction, it may check
+       * the output space for ungrouping variables. However if an
+       * ungrouping variable is a measure, the field expression value
+       * needed to get the definition needs to be computed in the
+       * input space of the query. There is a test which failed which
+       * caused this code to be here, but this is really a bandaid.
+       *
+       * Some re-work of how to get the definition of a SpaceField
+       * no matter what it is contained in needs to be thought out.
+       *
+       * ... or this check would look at the finalized output of
+       * the namespace and not re-compile ...
+       */
+      const definition = found.fieldDef();
+      if (definition) {
+        head.addReference({
+          type:
+            found instanceof StructSpaceField
+              ? "joinReference"
+              : "fieldReference",
+          definition,
+          location: head.location,
+          text: head.refString,
+        });
+      }
+    }
+    if (rest.length) {
+      if (found instanceof StructSpaceField) {
+        return found.fieldSpace.lookup(rest);
+      }
+      return {
+        error: `'${head}' cannot contain a '${rest[0]}'`,
+        found: undefined,
+      };
+    }
+    return { found, error: undefined };
+  }
+}
+
+/**
+ * A FieldSpace which may undergo modification
+ */
+export class DynamicSpace extends StaticSpace {
+  protected final: model.StructDef | undefined;
+  protected source: SpaceSeed;
+  outputFS?: QuerySpace;
+  completions: (() => void)[] = [];
+  private complete = false;
+
+  constructor(extending: SourceSpec) {
+    const source = new SpaceSeed(extending);
+    super(cloneDeep(source.structDef));
+    this.final = undefined;
+    this.source = source;
+  }
+
+  /**
+   * Factory for DynamicSpace when there are accept/except edits
+   * @param from A structdef which seeds this space
+   * @param choose A accept/except edit of the "from" fields
+   */
+  static filteredFrom(
+    from: model.StructDef,
+    choose?: FieldListEdit
+  ): DynamicSpace {
+    const edited = new DynamicSpace(from);
+    if (choose) {
+      const names = choose.refs.list;
+      const oldMap = edited.entries();
+      edited.dropEntries();
+      for (const [symbol, value] of oldMap) {
+        const included = !!names.find((f) => f.refString === symbol);
+        const accepting = choose.edit === "accept";
+        if (included === accepting) {
+          edited.setEntry(symbol, value);
+        }
+      }
+    }
+    return edited;
+  }
+
+  whenComplete(finalizeStep: () => void): void {
+    if (this.complete) {
+      finalizeStep();
+    } else {
+      this.completions.push(finalizeStep);
+    }
+  }
+
+  isComplete(): void {
+    this.complete = true;
+    for (const step of this.completions) {
+      step();
+    }
+    this.completions = [];
+  }
+
+  protected setEntry(name: string, value: SpaceEntry): void {
+    if (this.final) {
+      throw new Error("Space already final");
+    }
+    super.setEntry(name, value);
+  }
+
+  addParameters(params: HasParameter[]): DynamicSpace {
+    for (const oneP of params) {
+      this.setEntry(oneP.name, new AbstractParameter(oneP));
+    }
+    return this;
+  }
+
+  newEntry(name: string, astEl: MalloyElement, entry: SpaceEntry): void {
+    if (this.entry(name)) {
+      astEl.log(`Cannot redefine '${name}'`);
+      return;
+    }
+    this.setEntry(name, entry);
+  }
+
+  addField(...defs: ExploreField[]): void {
+    for (const def of defs) {
+      // TODO express the "three fields kinds" in a typesafe way
+      // one of three kinds of fields are legal in an explore: expressions ...
+      const elseLog = def.log;
+      const elseType = def.elementType;
+      if (def instanceof FieldDeclaration) {
+        const exprField = new ExpressionFieldFromAst(this, def);
+        this.newEntry(exprField.name, def, exprField);
+        // querry (turtle) fields
+      } else if (def instanceof TurtleDecl) {
+        const name = def.name;
+        this.newEntry(name, def, new QueryFieldAST(this, def, name));
+      } else if (def instanceof RenameField) {
+        if (def.oldName.refString === def.newName) {
+          def.log("Can't rename field to itself");
+          continue;
+        }
+        const oldValue = def.oldName.getField(this);
+        if (oldValue.found) {
+          if (oldValue.found instanceof SpaceField) {
+            this.setEntry(
+              def.newName,
+              new RenameSpaceField(oldValue.found, def.newName, def.location)
+            );
+            this.dropEntry(def.oldName.refString);
+          } else {
+            def.log(`'${def.oldName}' cannot be renamed`);
+          }
+        } else {
+          def.log(`Can't rename '${def.oldName}', no such field`);
+        }
+      } else if (def instanceof Join) {
+        this.newEntry(def.name.refString, def, new JoinSpaceField(this, def));
+      } else {
+        elseLog(
+          `Internal error: Expected expression, query, or rename, got '${elseType}'`
+        );
+      }
+    }
+  }
+
+  addFieldDef(fd: model.FieldDef): void {
+    this.setEntry(nameOf(fd), this.fromFieldDef(fd));
+  }
+
+  structDef(): model.StructDef {
+    const parameters = this.fromStruct.parameters || {};
+    if (this.final === undefined) {
+      this.final = {
+        ...this.fromStruct,
+        fields: [],
+      };
+      // Need to process the entities in specific order
+      const fields: [string, SpaceField][] = [];
+      const joins: [string, SpaceField][] = [];
+      const turtles: [string, SpaceField][] = [];
+      const fixupJoins: [Join, model.StructDef][] = [];
+      for (const [name, spaceEntry] of this.entries()) {
+        if (spaceEntry instanceof StructSpaceField) {
+          joins.push([name, spaceEntry]);
+        } else if (spaceEntry instanceof QueryField) {
+          turtles.push([name, spaceEntry]);
+        } else if (spaceEntry instanceof SpaceField) {
+          fields.push([name, spaceEntry]);
+        } else if (spaceEntry instanceof SpaceParam) {
+          parameters[name] = spaceEntry.parameter();
+        }
+      }
+      const reorderFields = [...fields, ...joins, ...turtles];
+      for (const [fieldName, field] of reorderFields) {
+        if (field instanceof JoinSpaceField) {
+          const joinStruct = field.join.structDef();
+          if (!ErrorFactory.isErrorStructDef(joinStruct)) {
+            this.final.fields.push(joinStruct);
+            fixupJoins.push([field.join, joinStruct]);
+          }
+        } else {
+          const fieldDef = field.fieldDef();
+          if (fieldDef) {
+            this.final.fields.push(fieldDef);
+          } else {
+            throw new Error(`'${fieldName}' doesn't have a FieldDef`);
+          }
+        }
+      }
+      if (Object.entries(parameters).length > 0) {
+        this.final.parameters = parameters;
+      }
+      // If we have join expressions, we need to now go back and fill them in
+      for (const [join, missingOn] of fixupJoins) {
+        join.fixupJoinOn(this, missingOn);
+      }
+    }
+    this.isComplete();
+    return this.final;
+  }
+}
+
+export abstract class ResultSpace extends DynamicSpace {
+  readonly exprSpace: QuerySpace;
+  astEl?: MalloyElement | undefined;
+  abstract readonly segmentType: "reduce" | "project" | "index";
+  constructor(readonly queryInputSpace: FieldSpace) {
+    super(queryInputSpace.emptyStructDef());
+    this.exprSpace = new QuerySpace(queryInputSpace, this);
+  }
+
+  log(s: string): void {
+    if (this.astEl) {
+      this.astEl.log(s);
+    }
+  }
+
+  addMembers(members: FieldCollectionMember[]): void {
+    for (const member of members) {
+      if (member instanceof FieldReference) {
+        this.addReference(member);
+      } else if (member instanceof WildcardFieldReference) {
+        this.addWild(member);
+      } else {
+        this.addField(member);
+      }
+    }
+  }
+
+  addReference(ref: FieldReference): void {
+    const refName = ref.outputName;
+    if (this.entry(refName)) {
+      ref.log(`Output already has a field named '${refName}'`);
+    } else {
+      this.setEntry(refName, new ReferenceField(ref));
+    }
+  }
+
+  addWild(wild: WildcardFieldReference): void {
+    this.setEntry(wild.refString, new WildSpaceField(wild.refString));
+  }
+
+  /**
+   * Check for the definition of an ungrouping reference in the result space,
+   * or in the case of an exclude reference, if this query is nested
+   * in another query, in the result space of a query that this query
+   * is nested inside of.
+   */
+  checkUngroup(fn: FieldName, isExclude: boolean): void {
+    if (!this.entry(fn.refString)) {
+      if (isExclude && this.exprSpace.nestParent) {
+        const parent = this.exprSpace.nestParent;
+        parent.whenComplete(() => {
+          parent.result.checkUngroup(fn, isExclude);
+        });
+      } else {
+        const uName = isExclude ? "exclude()" : "all()";
+        fn.log(`${uName} '${fn.refString}' is missing from query output`);
+      }
+    }
+  }
+
+  addQueryItems(...qiList: QueryItem[]): void {
+    for (const qi of qiList) {
+      if (qi instanceof FieldReference || qi instanceof NestReference) {
+        this.addReference(qi);
+      } else if (qi instanceof FieldDeclaration) {
+        this.addField(qi);
+      } else if (isNestedQuery(qi)) {
+        const qf = new QueryFieldAST(this, qi, qi.name);
+        qf.nestParent = this.exprSpace;
+        this.setEntry(qi.name, qf);
+      } else {
+        // Compiler will error if we don't handle all cases
+        const _itemNotHandled: never = qi;
+      }
+    }
+  }
+
+  canContain(_qd: model.QueryFieldDef): boolean {
+    return true;
+  }
+
+  protected queryFieldDefs(): model.QueryFieldDef[] {
+    const fields: model.QueryFieldDef[] = [];
+    for (const [name, field] of this.entries()) {
+      if (field instanceof SpaceField) {
+        const fieldQueryDef = field.getQueryFieldDef(this.exprSpace);
+        if (fieldQueryDef) {
+          if (this.canContain(fieldQueryDef)) {
+            fields.push(fieldQueryDef);
+          } else {
+            this.log(`'${name}' not legal in ${this.segmentType}`);
+          }
+        } else {
+          throw new Error(`'${name}' does not have a QueryFieldDef`);
+        }
+      }
+    }
+    this.isComplete();
+    return fields;
+  }
+
+  getQuerySegment(rf: model.QuerySegment | undefined): model.QuerySegment {
+    const p = this.getPipeSegment(rf);
+    if (model.isQuerySegment(p)) {
+      return p;
+    }
+    throw new Error("TODO NOT POSSIBLE");
+  }
+
+  getPipeSegment(
+    refineFrom: model.QuerySegment | undefined
+  ): model.PipeSegment {
+    if (this.segmentType == "index") {
+      // TODO ... should make this go away
+      throw new Error("INDEX FIELD PIPE SEGMENT MIS HANDLED");
+    }
+
+    if (refineFrom?.extendSource) {
+      for (const xField of refineFrom.extendSource) {
+        this.exprSpace.addFieldDef(xField);
+      }
+    }
+
+    const segment: model.QuerySegment = {
+      type: this.segmentType,
+      fields: this.queryFieldDefs(),
+    };
+
+    segment.fields = mergeFields(refineFrom?.fields, segment.fields);
+
+    if (refineFrom?.extendSource) {
+      segment.extendSource = refineFrom.extendSource;
+    }
+    if (this.exprSpace.extendList.length > 0) {
+      const newExtends: model.FieldDef[] = [];
+      const extendedStruct = this.exprSpace.structDef();
+
+      for (const extendName of this.exprSpace.extendList) {
+        const extendEnt = extendedStruct.fields.find(
+          (f) => nameOf(f) == extendName
+        );
+        if (extendEnt) {
+          newExtends.push(extendEnt);
+        }
+      }
+      segment.extendSource = mergeFields(segment.extendSource, newExtends);
+    }
+    this.isComplete();
+    return segment;
+  }
+}
+
+/**
+ * Reduce and project queries use a "QuerySpace"
+ */
+export class ReduceFieldSpace extends ResultSpace {
+  readonly segmentType = "reduce";
+}
+
+export class ProjectFieldSpace extends ResultSpace {
+  readonly segmentType = "project";
+
+  canContain(qd: model.QueryFieldDef): boolean {
+    if (typeof qd !== "string") {
+      if (model.isFilteredAliasedName(qd)) {
+        return true;
+      }
+      if (qd.type === "turtle") {
+        this.log("Cannot nest queries in project");
+        return false;
+      }
+      if (model.expressionIsAggregate(qd.expressionType)) {
+        this.log("Cannot add aggregate measures to project");
+        return false;
+      }
+    }
+    return true;
+  }
+}
+
+export abstract class SpaceEntry {
+  abstract type(): FieldType;
+  abstract refType: "field" | "parameter";
+}
+type FieldMap = Record<string, SpaceEntry>;
+
+export abstract class SpaceField extends SpaceEntry {
+  readonly refType = "field";
+
+  protected fieldTypeFromFieldDef(def: model.FieldDef): FieldType {
+    const ref: FieldType = { type: def.type };
+    if (model.isFieldTypeDef(def) && def.expressionType) {
+      ref.expressionType = def.expressionType;
+    }
+    return ref;
+  }
+
+  getQueryFieldDef(_fs: FieldSpace): model.QueryFieldDef | undefined {
+    return undefined;
+  }
+
+  fieldDef(): model.FieldDef | undefined {
+    return undefined;
+  }
+}
+
+export interface FieldType {
+  type: FieldValueType;
+  expressionType?: model.ExpressionType;
+}
+
+export class StructSpaceField extends SpaceField {
+  protected space?: FieldSpace;
+  constructor(protected sourceDef: model.StructDef) {
+    super();
+  }
+
+  get fieldSpace(): FieldSpace {
+    if (!this.space) {
+      this.space = new StaticSpace(this.sourceDef);
+    }
+    return this.space;
+  }
+
+  fieldDef(): model.FieldDef {
+    return this.sourceDef;
+  }
+
+  type(): FieldType {
+    return { type: "struct" };
+  }
+}
+
+export abstract class QueryField extends SpaceField {
+  constructor(protected inSpace: FieldSpace) {
+    super();
+  }
+
+  abstract getQueryFieldDef(fs: FieldSpace): model.QueryFieldDef | undefined;
+  abstract fieldDef(): model.FieldDef;
+
+  type(): FieldType {
+    return { type: "turtle" };
+  }
+}
+
+export class QueryFieldStruct extends QueryField {
+  constructor(fs: FieldSpace, protected turtleDef: model.TurtleDef) {
+    super(fs);
+  }
+
+  rename(name: string): void {
+    this.turtleDef = {
+      ...this.turtleDef,
+      as: name,
+    };
+  }
+
+  fieldDef(): model.TurtleDef {
+    return this.turtleDef;
+  }
+
+  getQueryFieldDef(_fs: FieldSpace): model.QueryFieldDef | undefined {
+    return this.fieldDef();
+  }
+}
+
+export class ColumnSpaceField extends SpaceField {
+  constructor(protected def: model.FieldTypeDef) {
+    super();
+  }
+
+  fieldDef(): model.FieldDef {
+    return this.def;
+  }
+
+  type(): FieldType {
+    return this.fieldTypeFromFieldDef(this.def);
+  }
+}
+
+export abstract class SpaceParam extends SpaceEntry {
+  abstract parameter(): model.Parameter;
+  readonly refType = "parameter";
+}
+
+export class DefinedParameter extends SpaceParam {
+  constructor(readonly paramDef: model.Parameter) {
+    super();
+  }
+
+  parameter(): model.Parameter {
+    return this.paramDef;
+  }
+
+  type(): FieldType {
+    return { type: this.paramDef.type };
+  }
+}
+
+/**
+ * Based on how things are constructed, the starting field space
+ * can either be another field space or an existing structdef.
+ * Using a SpaceSeed allows a class to accept either one
+ * and use either version at some future time.
+ */
+export class SpaceSeed {
+  private spaceSpec: SourceSpec;
+  private asFS?: StaticSpace;
+  private asStruct?: model.StructDef;
+
+  constructor(readonly sourceSpec: SourceSpec) {
+    this.spaceSpec = sourceSpec;
+  }
+
+  get structDef(): model.StructDef {
+    if (isFieldSpace(this.spaceSpec)) {
+      if (!this.asStruct) {
+        this.asStruct = this.spaceSpec.structDef();
+      }
+      return this.asStruct;
+    }
+    return this.spaceSpec;
+  }
+
+  get fieldSpace(): FieldSpace {
+    if (isFieldSpace(this.spaceSpec)) {
+      return this.spaceSpec;
+    }
+    if (!this.asFS) {
+      this.asFS = new StaticSpace(this.spaceSpec);
+    }
+    return this.asFS;
+  }
+}
+
+export type SourceSpec = model.StructDef | FieldSpace;
+function isFieldSpace(x: SourceSpec): x is FieldSpace {
+  return x.type == "fieldSpace";
+}
+
+export class AbstractParameter extends SpaceParam {
+  constructor(readonly astParam: HasParameter) {
+    super();
+  }
+
+  parameter(): model.Parameter {
+    return this.astParam.parameter();
+  }
+
+  type(): FieldType {
+    const type = this.astParam.type || "unknown";
+    return { type };
+  }
+}
+
+export class ExpressionFieldFromAst extends SpaceField {
+  fieldName: string;
+  defType?: FieldType;
+  constructor(
+    readonly space: DynamicSpace,
+    readonly exprDef: FieldDeclaration
+  ) {
+    super();
+    this.fieldName = exprDef.defineName;
+  }
+
+  get name(): string {
+    return this.fieldName;
+  }
+
+  fieldDef(): model.FieldDef {
+    const def = this.exprDef.fieldDef(this.space, this.name);
+    this.defType = this.fieldTypeFromFieldDef(def);
+    return def;
+  }
+
+  getQueryFieldDef(fs: FieldSpace): model.QueryFieldDef {
+    const def = this.exprDef.queryFieldDef(fs, this.name);
+    this.defType = this.fieldTypeFromFieldDef(def);
+    return def;
+  }
+
+  type(): FieldType {
+    if (this.defType) {
+      return this.defType;
+    }
+    return this.fieldTypeFromFieldDef(this.fieldDef());
+  }
+}
+
+export class QueryFieldAST extends QueryField {
+  renameAs?: string;
+  nestParent?: QuerySpace;
+  constructor(
+    fs: FieldSpace,
+    readonly turtle: TurtleDecl,
+    protected name: string
+  ) {
+    super(fs);
+  }
+
+  getQueryFieldDef(fs: FieldSpace): model.QueryFieldDef {
+    const def = this.turtle.getFieldDef(fs, this.nestParent);
+    if (this.renameAs) {
+      def.as = this.renameAs;
+    }
+    return def;
+  }
+
+  fieldDef(): model.TurtleDef {
+    const def = this.turtle.getFieldDef(this.inSpace, this.nestParent);
+    if (this.renameAs) {
+      def.as = this.renameAs;
+    }
+    return def;
+  }
+}
+
+export class RenameSpaceField extends SpaceField {
+  constructor(
+    private readonly otherField: SpaceField,
+    private readonly newName: string,
+    private readonly location: model.DocumentLocation
+  ) {
+    super();
+  }
+
+  fieldDef(): model.FieldDef | undefined {
+    const renamedFieldRaw = this.otherField.fieldDef();
+    if (renamedFieldRaw === undefined) {
+      return undefined;
+    }
+    return {
+      ...renamedFieldRaw,
+      as: this.newName,
+      location: this.location,
+    };
+  }
+
+  type(): FieldType {
+    return this.otherField.type();
+  }
+}
+
+export class JoinSpaceField extends StructSpaceField {
+  constructor(readonly intoFS: FieldSpace, readonly join: Join) {
+    super(join.structDef());
+  }
+}
+
+/**
+ * Unlike a source, which is a refinement of a namespace, a query
+ * is creating a new unrelated namespace. The query starts with a
+ * source, which it might modify. This set of fields used to resolve
+ * expressions in the query is called the "input space". There is a
+ * specialized QuerySpace for each type of query operation.
+ *
+ * The query output is managed by an instance of ResultSpace.
+ */
+
+export class QuerySpace extends DynamicSpace {
+  nestParent?: QuerySpace;
+  extendList: string[] = [];
+
+  constructor(input: SourceSpec, readonly result: ResultSpace) {
+    const inputSpace = new SpaceSeed(input);
+    super(inputSpace.structDef);
+  }
+
+  extendSource(extendField: Join | FieldDeclaration): void {
+    this.addField(extendField);
+    if (extendField instanceof Join) {
+      this.extendList.push(extendField.name.refString);
+    } else {
+      this.extendList.push(extendField.defineName);
+    }
+  }
+}
+
+export class ReferenceField extends SpaceField {
+  constructor(readonly fieldRef: FieldReference) {
+    super();
+  }
+
+  getQueryFieldDef(fs: FieldSpace): model.QueryFieldDef | undefined {
+    // Lookup string and generate error if it isn't defined.
+    const check = this.fieldRef.getField(fs);
+    if (check.error) {
+      this.fieldRef.log(check.error);
+    }
+    return this.fieldRef.refString;
+  }
+
+  type(): FieldType {
+    return { type: "unknown" };
+  }
+}
+
+export class WildSpaceField extends SpaceField {
+  constructor(readonly wildText: string) {
+    super();
+  }
+
+  type(): FieldType {
+    throw new Error("should never ask a wild field for its type");
+  }
+
+  getQueryFieldDef(_fs: FieldSpace): model.QueryFieldDef {
+    return this.wildText;
+  }
+}
+
+interface LookupFound {
+  found: SpaceEntry;
+  error: undefined;
+}
+interface LookupError {
+  error: string;
+  found: undefined;
+}
+export type LookupResult = LookupFound | LookupError;
+
+export class IndexFieldSpace extends ResultSpace {
+  readonly segmentType = "index";
+  fieldList = new Set<string>();
+
+  addReference(ref: FieldReference): void {
+    if (ref.getField(this.exprSpace).found) {
+      this.fieldList.add(ref.refString);
+    }
+  }
+
+  addWild(wild: WildcardFieldReference): void {
+    this.fieldList.add(wild.refString);
+  }
+
+  getPipeSegment(refineIndex?: model.PipeSegment): model.IndexSegment {
+    if (refineIndex && refineIndex.fields) {
+      for (const exists of refineIndex.fields) {
+        if (typeof exists == "string") {
+          this.fieldList.add(exists);
+        }
+      }
+    }
+    this.isComplete();
+    return {
+      type: "index",
+      fields: Array.from(this.fieldList.values()),
+    };
   }
 }
